requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cffi==1.17.1
click==8.2.0
cryptography==44.0.3
fastadmin==0.2.22
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
mysqlclient==2.2.7
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2